moved CosmosElementSerializer out side of the CosmosElement namespace
