+++ Microsoft.Azure.Cosmos/src/Query/v3Query/ReadFeedResponse.cs
@@ -8,6 +8,7 @@ namespace Microsoft.Azure.Cosmos
     using System.Net;
     using Microsoft.Azure.Cosmos.CosmosElements;
     using Microsoft.Azure.Cosmos.Json;
+    using Microsoft.Azure.Cosmos.Serializer;
 
     internal class ReadFeedResponse<T> : FeedResponse<T>
     {
